+++ plugins/decl_hdf5/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
+* Support for reading and writing attributes of datasets & datagroups
+* Getting dataset size before reading it
 
 ### Changed